Added Float compatibility

+++ project/drivers/GUI_XYGraph.h
@@ -11,15 +11,15 @@
 #define AXIS_X_DATA_POINTS 80
 
 struct XYGraphData {
-    uint16_t prevDataX;
-    uint16_t prevDataY;
-    uint16_t data[AXIS_X_DATA_POINTS];
+    float prevDataX;
+    float prevDataY;
+    float data[AXIS_X_DATA_POINTS];
     // uint16_t y_max;
-    uint16_t y_estop;
+    float y_estop;
 
     uint16_t graphHead;
 
-    uint16_t densitySum;
+    float densitySum;
     uint16_t densityCount;
     uint16_t density;
     bool updateFlag;
@@ -41,7 +41,7 @@ extern void XYGraph_init_display(struct XYGraphFrame* frame, char* units);
 
 extern void GraphFrame_init(struct XYGraphFrame *frame, uint16_t x, uint16_t y, uint16_t w, uint16_t h);
 
-extern void GraphData_init(struct XYGraphData *graph, uint16_t density, uint16_t estop);
+extern void GraphData_init(struct XYGraphData *graph, uint16_t density, float estop);
 
 extern void drawGraphFrame(struct XYGraphFrame *frame);
 
@@ -59,7 +59,7 @@ extern void updateGraph(struct XYGraphFrame *frame, struct XYGraphData *graph);
 
 extern void drawGraphLag(struct XYGraphFrame* frame, UInt32 time);
 
-extern bool accumulateGraphData(struct XYGraphData *graph, uint16_t newData);
+extern bool accumulateGraphData(struct XYGraphData *graph, float newData);
 
 
 #endif /* DRIVERS_GUI_XYGRAPH_H_ */
